requirements.txt
dash
dash_core_components
dash_html_components
dash_bootstrap_components
plotly
datetime
pandas
numpy
psycopg2-binary
pyyaml
SQLAlchemy
flask